+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/domain/model/ToDoTask.kt
@@ -32,7 +32,7 @@ data class ToDoTask(
         const val TABLE_NAME = "todo_task_tab"
     }
 
-    fun dueLocalDateTime(): LocalDateTime =
+    fun dueLocalDateTime(): LocalDateTime? =
         LocalDateTime.ofInstant(
             Instant.ofEpochMilli(dueTimeInMillis),
             ZoneId.systemDefault()