+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/util/UiEvent.kt
@@ -1,7 +1,10 @@
 package com.gmail.matikano9.todoapp.util
 
 import androidx.navigation.compose.ComposeNavigator
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.spec.DestinationSpec
 import com.ramcosta.composedestinations.spec.Direction
+import com.ramcosta.composedestinations.spec.DirectionDestinationSpec
 
 sealed class UiEvent{
     object PopBackStack: UiEvent()